requirements.txt
aiohttp==2.3.6
aniso8601==1.3.0
asn1crypto==0.23.0
astroid==1.5.3
async-timeout==2.0.0
base58==0.2.5
BigchainDB==1.0.1
bigchaindb-driver==0.4.1
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
cryptoconditions==0.6.0.dev1
cryptography==1.9
Django==2.0
django-searchable-select==1.4.5
Flask==0.12.2
Flask-Cors==3.0.3
Flask-RESTful==0.3.6
gunicorn==19.7.1
idna==2.6
isort==4.2.15
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.5.1
lazy-object-proxy==1.3.1
logstats==0.2.1
MarkupSafe==1.0
mccabe==0.6.1
multidict==3.3.2
multipipes==0.1.0
pyasn1==0.2.3
pycparser==2.18
pylint==1.7.5
pymongo==3.6.0
PyNaCl==1.1.2
pysha3==1.0.2
python-dateutil==2.6.1
python-rapidjson==0.0.11
python-rapidjson-schema==0.1.1
pytz==2017.3
PyYAML==3.12
requests==2.18.4
rethinkdb==2.3.0.post6
six==1.11.0
statsd==3.2.1
urllib3==1.22
Werkzeug==0.13
wrapt==1.10.11
yarl==0.16.0
